fastapi-proxy/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# HTTP client
httpx==0.26.0

# JWT authentication
PyJWT[crypto]==2.8.0
cryptography==42.0.2

# Configuration and validation
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Logging and monitoring (optional but recommended)
python-json-logger==2.0.7
